cockpit.main(): show GUI with exc caught during initialisation (issue #439)
There are some initialisation steps before CockpitApp.OnInit.  If an
exception occurs there, then cockpit exits with errors on the command
line which the user does not see.  But there's no wx.App create yet to
display a dialog with the exception.  This change will create a simple
wx.App with the single purpose of displaying the exception traceback.